requirements.txt
azure-common==1.1.27
azure-core==1.18.0
azure-identity==1.6.1
azure-keyvault==4.1.0
azure-keyvault-certificates==4.3.0
azure-keyvault-keys==4.4.0
azure-keyvault-secrets==4.3.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
idna==3.2
isodate==0.6.0
msal==1.14.0
msal-extensions==0.3.0
msrest==0.6.21
oauthlib==3.1.1
portalocker==1.7.1
pycparser==2.20
PyJWT==2.1.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7
